[libc] Add iswctype and wctype (v2)
- Change wctype_t from unsigned int to unsigned short
- Move iswctype and wctype out of the ascii namespace
- Use wctype_t instead of raw integer types in wctype_utils.h
- Introduce a mapping table to bind property names to descriptors
- Fix header formatting issues

+++ libc/include/llvm-libc-types/wctype_t.h
@@ -1,5 +1,4 @@
-//===-- Definition of wctype_t types
-//----------------------------------------===//
+//===-- Definition of wctype_t types --------------------------------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -10,6 +9,6 @@
 #ifndef LLVM_LIBC_TYPES_WCTYPE_T_H
 #define LLVM_LIBC_TYPES_WCTYPE_T_H
 
-typedef unsigned int wctype_t;
+typedef unsigned short wctype_t;
 
 #endif // LLVM_LIBC_TYPES_WCTYPE_T_H
